Move shared into candlelib and tests into separate directory
