+++ candlelib/include/candle/MD/MD_strings.hpp
@@ -3,8 +3,8 @@
 #include <cctype>
 #include <cstdlib>
 #include <string>
-#include <mab_types.hpp>
-#include "md_types.hpp"
+#include <candle/types/mab_types.hpp>
+#include "candle/types/md_types.hpp"
 #include <map>
 #include <optional>
