+++ candlelib/include/candle/candlelib.hpp
@@ -6,7 +6,7 @@ It provides all modules and API interfaces.
 
 #include <unordered_map>
 
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include "candle/communication_device/candle.hpp"
 #include "candle/MD/MD.hpp"
 #include "candle/pds/pds.hpp"
