+++ candlelib/include/candle/communication_device/candle.hpp
@@ -16,9 +16,9 @@
 #include "candle/communication_interface/I_communication_interface.hpp"
 #include "candle/communication_interface/USB.hpp"
 #include "candle/communication_interface/SPI.hpp"
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include "candle/communication_device/candle_frame_adapter.hpp"
-#include "candle_frame_dto.hpp"
+#include "candle/types/candle_frame_dto.hpp"
 
 namespace mab
 {
