+++ candlelib/include/candle/communication_device/candle_frame_adapter.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "logger.hpp"
-#include "candle_frame_dto.hpp"
-#include "mab_types.hpp"
+#include "candle/types/candle_frame_dto.hpp"
+#include "candle/types/mab_types.hpp"
 
 #include <atomic>
 #include <array>
