+++ candlelib/include/candle/communication_interface/I_communication_interface.hpp
@@ -5,7 +5,7 @@
 #include <utility>
 #include <vector>
 
-#include <mab_types.hpp>
+#include <candle/types/mab_types.hpp>
 
 namespace mab
 {
