+++ candlelib/include/candle/communication_interface/SPI.hpp
@@ -11,7 +11,7 @@
 #include <string>
 #include <thread>
 
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include "candle/communication_interface/I_communication_interface.hpp"
 #include "logger.hpp"
 #include "candle/crc.hpp"
