+++ candlelib/include/candle/communication_interface/USB.hpp
@@ -7,7 +7,7 @@
 #include <mutex>
 #include <utility>
 
-#include <mab_types.hpp>
+#include <candle/types/mab_types.hpp>
 #include <logger.hpp>
 #include "candle/communication_interface/I_communication_interface.hpp"
