+++ candlelib/include/candle/objectDictionary/edsParser.hpp
@@ -10,7 +10,7 @@
 #include <algorithm>
 #include <cctype>
 #include "configHelpers.hpp"
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include "logger.hpp"
 #include "candle/objectDictionary/edsEntry.hpp"
 #include "mini/ini.h"
