+++ candlelib/include/candle/pds/pds_module.hpp
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include "logger.hpp"
 
-#include "pds_types.hpp"
+#include "candle/types/pds_types.hpp"
 #include "candle/communication_device/candle_types.hpp"
 #include "candle/pds/pds_protocol.hpp"
 #include "candle/communication_device/candle.hpp"
