+++ candlelib/include/candle/pds/pds_protocol.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
 #include <stdint.h>
-#include "pds_types.hpp"
+#include "candle/types/pds_types.hpp"
 #include "logger.hpp"
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include <string.h>
 #include <vector>
