+++ candlelib/include/candle/types/candle_frame_dto.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#include "mab_def.hpp"
-#include "mab_types.hpp"
+#include "candle/types/mab_def.hpp"
+#include "candle/types/mab_types.hpp"
 #include <cstring>
 #include <type_traits>
 namespace mab
