+++ candlelib/include/candle/types/mab_def.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include <stddef.h>
 
 namespace mab
