+++ candlelib/include/candle/types/manufacturer_data.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #ifdef __cplusplus
 #include <cstring>
 #else
