+++ candlelib/include/candle/types/pds_properties.hpp
@@ -2,7 +2,7 @@
 
 #include <array>
 #include <cstddef>
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 
 namespace mab
 {
