+++ candlelib/src/MD/MDCO.cpp
@@ -9,7 +9,7 @@
 #include "candle/MD/MDStatus.hpp"
 #include "candle/communication_device/candle_types.hpp"
 #include "candle/objectDictionary/edsEntry.hpp"
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 
 namespace mab
 {
