+++ candlelib/src/pds/pds.cpp
@@ -1,5 +1,5 @@
 #include "candle/pds/pds.hpp"
-#include "pds_types.hpp"
+#include "candle/types/pds_types.hpp"
 
 namespace mab
 {
