+++ candlelib/test/communication_device/candle_bootloader_test.cpp
@@ -1,6 +1,6 @@
 #include "candle/communication_interface/I_communication_interface_mock.hpp"
 
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include "logger.hpp"
 #include "candle/communication_device/candle_bootloader.hpp"
