+++ candletool/include/configHelpers.hpp
@@ -3,7 +3,7 @@
 #include <string>
 #include <filesystem>
 #include <optional>
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include "mini/ini.h"
 
 #ifdef WIN32
