+++ candletool/include/mabFileParser.hpp
@@ -6,7 +6,7 @@
 #include <memory>
 #include "logger.hpp"
 
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 
 class MabFileParser
 {
