+++ candletool/src/md_cli.cpp
@@ -11,8 +11,8 @@
 #include "canLoader.hpp"
 #include "candle/communication_device/candle.hpp"
 #include "logger.hpp"
-#include "mab_types.hpp"
-#include "md_types.hpp"
+#include "candle/types/mab_types.hpp"
+#include "candle/types/md_types.hpp"
 #include "candle/communication_device/candle_types.hpp"
 #include "mabFileParser.hpp"
 #include "md_cfg_map.hpp"
