+++ candletool/src/pds_cli.cpp
@@ -1,12 +1,12 @@
 #include "canLoader.hpp"
 #include "candle/communication_device/candle.hpp"
 #include "mabFileParser.hpp"
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include "candle/pds/pds.hpp"
 #include "pds_cli.hpp"
-#include "mab_def.hpp"
+#include "candle/types/mab_def.hpp"
 #include "configHelpers.hpp"
-#include "pds_types.hpp"
+#include "candle/types/pds_types.hpp"
 
 /*
     PDS Ini fields keywords
