+++ candletool/src/web_file_module/curl_handler.hpp
@@ -3,7 +3,7 @@
 
 #include "mini/ini.h"
 
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include "logger.hpp"
 #include "web_file.hpp"
