+++ candletool/src/web_file_module/flasher.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "web_file.hpp"
-#include "mab_types.hpp"
+#include "candle/types/mab_types.hpp"
 #include "utilities.hpp"
 #include "logger.hpp"
