+++ commons/CMakeLists.txt
@@ -1,10 +1,11 @@
 cmake_minimum_required(VERSION 3.8)
 project(commons)
 
-add_subdirectory(shared_data)
+include(GNUInstallDirs)
 
 add_library(logger STATIC logger/logger.cpp)
-target_include_directories(logger PUBLIC logger)
+target_include_directories(logger PUBLIC
+  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/logger>)
 target_compile_options(logger PRIVATE -fPIC)
 
 set(INSTALL_GTEST OFF)
@@ -14,8 +15,4 @@ add_subdirectory(unit_test_runner)
 if(MAKE_TESTS)
   add_unit_test_executable(logger_test logger/logger_test.cpp)
   target_link_libraries(logger_test PUBLIC logger)
-
-  add_unit_test_executable(candle_frame_dto_test
-                           candle_frame_dto_test/candle_frame_dto_test.cpp)
-  target_link_libraries(candle_frame_dto_test PUBLIC shared_data)
 endif()
